Split libtracepoint-decode from libeventheader-decode
Split a new `libtracepoint-decode-cpp` library out from the existing
`libeventheader-decode-cpp` library. This has the perf.data parser, the
metadata manager, and the event structures. In other words:

- `libtracepoint-decode-cpp` = perf.data parser, metadata, event structures.
- `libeventheader-decode-cpp` = EventEnumerator, EventFormatter, and
  perf-decode tool (since the perf-decode tool decodes EventHeader events).

`libtracepoint-decode-cpp`:
- LIB: `libtracepoint-decode`
- `#include <tracepoint/PerfByteReader.h>`
- `#include <tracepoint/PerfDataAbi.h>`
- `#include <tracepoint/PerfDataFile.h>`
- `#include <tracepoint/PerfEventInfo.h>`
- `#include <tracepoint/PerfEventMetadata.h>`
