+++ libeventheader-decode-cpp/tools/decode-perf.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <PerfDataDecode/PerfEventInfo.h>
-#include <PerfDataDecode/PerfDataFile.h>
-#include <PerfDataDecode/PerfDataAbi.h>
+#include <tracepoint/PerfEventInfo.h>
+#include <tracepoint/PerfDataFile.h>
+#include <tracepoint/PerfDataAbi.h>
 #include <eventheader/EventFormatter.h>
 
 #include <string.h>
