+++ libtracepoint-decode-cpp/CMakeLists.txt
@@ -1,21 +1,11 @@
 cmake_minimum_required(VERSION 3.10)
 project(eventheader-decode-cpp
     VERSION 1.0.0
-    DESCRIPTION "EventHeader tracepoint decoding for C/C++"
+    DESCRIPTION "Tracepoint decoding for C/C++"
     HOMEPAGE_URL "https://github.com/microsoft/LinuxTracepoints"
     LANGUAGES CXX)
 include(GNUInstallDirs)
 include(CMakePackageConfigHelpers)
-set(BUILD_SAMPLES ON CACHE BOOL "Build sample code")
-set(BUILD_TOOLS ON CACHE BOOL "Build tool code")
-
-if(NOT TARGET tracepoint-decode)
-    find_package(tracepoint-decode 1.0 REQUIRED)
-endif()
-
-if(NOT TARGET eventheader-headers)
-    find_package(eventheader-headers 1.0 REQUIRED)
-endif()
 
 if(WIN32)
     add_compile_options(/W4 /WX /permissive-)
@@ -34,11 +24,3 @@ else()
 endif()
 
 add_subdirectory(src)
-
-if(BUILD_SAMPLES)
-    add_subdirectory(samples)
-endif()
-
-if(BUILD_TOOLS)
-    add_subdirectory(tools)
-endif()
