+++ libtracepoint-decode-cpp/src/PerfByteReader.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <PerfDataDecode/PerfByteReader.h>
+#include <tracepoint/PerfByteReader.h>
 #include <string.h>
 #include <assert.h>
