+++ libtracepoint-decode-cpp/src/PerfEventMetadata.cpp
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <PerfDataDecode/PerfEventMetadata.h>
-#include <PerfDataDecode/PerfByteReader.h>
+#include <tracepoint/PerfEventMetadata.h>
+#include <tracepoint/PerfByteReader.h>
 #include <assert.h>
 #include <errno.h>
 #include <stdlib.h>
