+++ libtracepoint-decode-cpp/src/tracepoint-decodeConfig.cmake.in
@@ -1,0 +1,2 @@
+@PACKAGE_INIT@
+include("${CMAKE_CURRENT_LIST_DIR}/tracepoint-decodeTargets.cmake")
